fix(protocol): widen frame header frameOff to u32 (#46)
Widen frameOff from u16 to u32: offsets within a 1 MiB block reach ~1 MiB, so they wrap past u16's 64 KiB bound and falsely signal a fresh block mid-block, corrupting reassembly. The 16-byte header is preserved by consuming the trailing reserved field, updated in lockstep across the TS and Go codecs and the shared AEAD vector.

+++ packages/wire/aead.go
@@ -17,7 +17,7 @@ type FrameHeaderInput struct {
 	Flags    uint8
 	FileIdx  uint16
 	BlockIdx uint32
-	FrameOff uint16
+	FrameOff uint32
 }
 
 // nonce builds the 12-byte GCM nonce: the direction's 4-byte salt followed by a
